fix(temporary-fix): added some sub-dependencies needed internally to fix expo build error
Fixed the "expo.core.ExpoModulesPackage" related import error by installing the "expo-modules-autolinking" package manually in the package.json for a temporary fix. Currently, that is a bug in expo itself leading to autolinking issues due it referring to the old "ExpoModulesPackage" package instead of modern equivalents. We wait for official patch for this bug. This fix will work till then.